+++ package.json
@@ -22,7 +22,6 @@
     "immutable": "^3.8.1",
     "lodash": "^4.15.0",
     "mammoth": "^1.2.1",
-    "mv": "^2.1.1",
     "mz": "^2.4.0",
     "ncp": "^2.0.0",
     "progress": "^1.1.8",
